# MAINT: special: performance optimization for simple autodifferentiation