@izaid 2024-09-17
This is great! Yes, @steppi and I discussed this, and there were some very simple optimisations to make that apparently brought down the runtime significantly.

It's actually quite valuable to see that table. I'm actually not surprised that the unoptimised autograd for `diff_n = 0` is faster than before, it used to wrap it in a tuple for no reason.

I also want to point out that the non-autograd `diff_n = 2` implementations have only existed for two months when I put them in. :) But good to see we are competitive with my own brutally hand-differentiated code.

Great work and thanks @steppi!