@fancidev 2024-09-17
Impressive speed-up!

Small question: it seems `dumb_binom` is only coded for `n <= 2`. Do you foresee a situation where larger `n` is needed?