@izaid 2024-09-17
> Impressive speed-up!
> 
> Small question: it seems `dumb_binom` is only coded for `n <= 2`. Do you foresee a situation where larger `n` is needed?

Theoretically, we should support any order of differentiation. I guess we'll replace dumb_binom at some point with the actual binom, but using a look-up table for small `n`.