@steppi 2024-09-23
I just pushed a commit to fix a misplaced `#endif` in `xsf/config.h` which broke things in CUDA (and thus `test_xsf_cuda.h`), but not anything public in SciPy.