@mdhaber 2024-10-04
The changes look good to me modulo https://github.com/scipy/scipy/pull/21568#discussion_r1788235586; however, that is really a separate issue, so I'll go ahead and merge this.